pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cogsci-skilllib"
version = "0.1.0a0"
description = "Standards-first cognitive-science agent skills scaffold."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "REPLACE_WITH_MAINTAINER_NAME" }
]
license = { text = "Provisional; choose final OSS license before public release." }
dependencies = [
  "PyYAML>=6.0",
  "jsonschema>=4.18",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
behavior = [
  "numpy>=2.0",
  "pandas>=2.2",
]
neuro = [
  "mne>=1.10",
  "mne-bids>=0.18",
]
modeling = [
  "pymc>=5.9",
  "arviz>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cogsci_skilllib = ["vendor/jspsych/*", "vendor/jspsych/**/*"]
